move config package and add provider token params
